+++ server/hubbedgame_manager.py
@@ -57,6 +57,57 @@ class _HubbedGameTrivialInherited(_GameWithAreas):
 
         return super().get_id()
 
+    def get_numerical_id(self) -> int:
+        """
+        Return the numerical portion of the ID of this hubbed game.
+
+        Returns
+        -------
+        int
+            Numerical portion of the ID.
+        """
+
+        return super().get_numerical_id()
+
+    def get_name(self) -> str:
+        """
+        Get the name of the hubbed game.
+
+        Returns
+        -------
+        str
+            Name.
+        """
+
+        return super().get_name()
+
+    def set_name(self, name: str):
+        """
+        Set the name of the hubbed game.
+
+        Parameters
+        ----------
+        name : str
+            Name.
+        """
+
+        self.unchecked_set_name(name)
+        self.manager._check_structure()
+
+    def unchecked_set_name(self, name: str):
+        """
+        Set the name of the hubbed game.
+
+        This method does not assert structural integrity.
+
+        Parameters
+        ----------
+        name : str
+            Name.
+        """
+
+        super().unchecked_set_name(name)
+
     def get_player_limit(self) -> Union[int, None]:
         """
         Return the player membership limit of this hubbed game.
@@ -2005,16 +2056,16 @@ class _HubbedGame(_HubbedGameTrivialInherited):
             })
         self.manager: HubbedGameManager  # Setting for typing
 
-    def get_name(self) -> str:
+    def get_type_name(self) -> str:
         """
-        Return the name of the hubbed game. Names are fully lowercase.
-        Implementations of the class should replace this with a human readable name of the game
-        with areas.
+        Return the type name of the hubbed game. Names are fully lowercase.
+        Implementations of the class should replace this with a human readable name of the hubbed
+        game.
 
         Returns
         -------
         str
-            Name of the hubbed game.
+            Type name of the hubbed game.
 
         """