+++ server/trial_manager.py
@@ -58,6 +58,57 @@ class _TrialTrivialInherited(_HubbedGame):
 
         return super().get_id()
 
+    def get_numerical_id(self) -> int:
+        """
+        Return the numerical portion of the ID of this trial.
+
+        Returns
+        -------
+        int
+            Numerical portion of the ID.
+        """
+
+        return super().get_numerical_id()
+
+    def get_name(self) -> str:
+        """
+        Get the name of the trial.
+
+        Returns
+        -------
+        str
+            Name.
+        """
+
+        return super().get_name()
+
+    def set_name(self, name: str):
+        """
+        Set the name of the trial.
+
+        Parameters
+        ----------
+        name : str
+            Name.
+        """
+
+        self.unchecked_set_name(name)
+        self.manager._check_structure()
+
+    def unchecked_set_name(self, name: str):
+        """
+        Set the name of the trial.
+
+        This method does not assert structural integrity.
+
+        Parameters
+        ----------
+        name : str
+            Name.
+        """
+
+        super().unchecked_set_name(name)
+
     def get_player_limit(self) -> Union[int, None]:
         """
         Return the player membership limit of this trial.
@@ -1767,15 +1818,15 @@ class _Trial(_TrialTrivialInherited):
             })
         self.manager: TrialManager  # Setting for typing
 
-    def get_name(self) -> str:
+    def get_type_name(self) -> str:
         """
-        Return the name of the trial. Names are fully lowercase.
+        Return the type name of the trial. Names are fully lowercase.
         Implementations of the class should replace this with a human readable name of the trial.
 
         Returns
         -------
         str
-            Name of the trial.
+            Type name of the trial.
 
         """
 
@@ -2911,7 +2962,7 @@ class _Trial(_TrialTrivialInherited):
                 # Check if client was added to any minigames automatically or not.
                 # If so, notifidy leaders
                 for minigame in self.get_minigames():
-                    name = minigame.get_name()
+                    name = minigame.get_type_name()
                     if minigame.is_player(client):
                         client.send_ooc(f'You were automatically added to {name} '
                                         f'`{minigame.get_id()}`.')